+++ bin/update.sh
@@ -48,16 +48,24 @@ while read -r line; do
 		tool="$(echo "${line}" | awk '{print $1}')"
 
 		debug "checking if ${tool} should be evaluated"
-		if [[ ${exclusions} =~ (^|,)" "*"${tool}"" "*($|,) ]]; then
-			info "skipping ${tool} because it is configured in the exclusion rule"
-			continue
+		if [[ -n ${exclusions} ]]; then
+			debug "checking if ${tool} is in the exclusion input"
+			if [[ ${exclusions} =~ (^|,)" "*"${tool}"" "*($|,) ]]; then
+				info "skipping ${tool} because it is in the exclusion input"
+				continue
+			fi
+		else
+			debug "no exclusions configured"
 		fi
 
 		if [[ -n ${inclusions} ]]; then
+			debug "checking if ${tool} is NOT in the inclusion input"
 			if [[ ! ${inclusions} =~ (^|,)" "*"${tool}"" "*($|,) ]]; then
-				info "skipping ${tool} because it is NOT configured in the inclusion rule"
+				info "skipping ${tool} because it is NOT in the inclusion input"
 				continue
 			fi
+		else
+			debug "no inclusions configured"
 		fi
 
 		info "evaluating ${tool}..."
@@ -70,26 +78,28 @@ while read -r line; do
 			info "update available for ${tool}"
 
 			if [[ -n ${skips} ]]; then
-				debug "checking if ${tool}@${latest_version} should be skipped..."
+				debug "checking if ${tool}@${latest_version} is in the skip input"
 				while read -r line; do
 					case "${line}" in
 					"")
-						debug "skipping empty line"
+						debug "skipping empty skip input line"
 						;;
 					*)
-						debug "processing skip line ('${line}')"
+						debug "processing skip input line ('${line}')"
 
 						name="$(echo "${line}" | awk '{print $1}')"
 						version="$(echo "${line}" | awk '{print $2}')"
 						debug "found skip mandate for ${name}@${version}"
 
 						if [ "${name}" == "${tool}" ] && [ "${version}" == "${latest_version}" ]; then
-							info "skipping ${name}@${version} because it is configured in the skip rule"
+							info "skipping ${name}@${version} because it is in the skip input"
 							continue 2
 						fi
 						;;
 					esac
 				done <<<"${skips}"
+			else
+				debug "no skips configured"
 			fi
 
 			debug "installing ${tool}@${latest_version}"